requirements.txt
google-api-core==1.22.1 
google-auth==1.20.1
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.52.0
graphviz==0.13.2
greenlet==0.4.16
grpcio==1.31.0
gTTS==2.1.1
gTTS-token==1.1.3